Refactor response handling with mixin
